requirements.txt
customtkinter
python-dotenv
requests